tool/backend/requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.11
filelock==3.17.0
fsspec==2025.3.0
h11==0.14.0
huggingface-hub==0.29.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
peft==0.15.0
pillow==11.1.0
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
pymupdf4llm==0.0.17
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
